pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "thanakan"
version = "0.3.13"
description = "Awesome `thanakan` is a Python cli/package created with https://github.com/TezRomacH/python-package-template"
readme = "README.md"
authors = [
  "codustry <hello@codustry.com>"
]
license = "MIT"
repository = "https://github.com/codustry/thanakan"
homepage = "https://github.com/codustry/thanakan"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]


[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.7.3"
crccheck = "^1.0"
httpx = {extras = ["http2"], version = "^0.16.1"}
parse-with-dot-access = "^1.18.0"
locate = "^0.0.1"
furl = "^2.1.0"
fastapi-utils = "^0.2.1"
httpx-auth = "^0.8.0"
loguru = "^0.5.3"
google-cloud-documentai = "^0.3.0"
Jinja2 = "^2.11.3"
Pillow = "^8.1.0"
pyzbar-x = "^0.2.1"

[tool.poetry.dev-dependencies]
darglint = "^1.5.4"
isort = "^5.5.4"
pyupgrade = "^2.7.2"
black = "^20.8b1"
mypy = "^0.800"
bandit = "^1.6.2"
safety = "^1.9.0"
pytest = "^6.0.1"
pylint = "^2.5.3"
pydocstyle = "^5.0.2"
pre-commit = "^2.7.1"
pytest-asyncio = "^0.14.0"
jupyterlab = "^3.0.8"

[tool.black]
# https://github.com/psf/black
line-length = 80
target-version = ["py37"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 80
